--- user/src/main/java/com/zerobase/user/controller/ReissueController.java
@@ -1,24 +1,0 @@
-package com.zerobase.user.controller;
-
-import com.zerobase.user.dto.response.ResponseMessage;
-import com.zerobase.user.service.ReissueService;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RestController
-public class ReissueController {
-
-    private final ReissueService reissueService;
-
-    @PostMapping("/reissue")
-    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response) {
-        reissueService.reissue(request, response);
-        return ResponseEntity.status(HttpStatus.OK).body(ResponseMessage.success());
-    }
-}
