+++ user/src/main/java/com/zerobase/user/jwt/LoginFilter.java
@@ -90,7 +90,7 @@ public class LoginFilter extends UsernamePasswordAuthenticationFilter {
         String role = auth.getAuthority();
 
         //토큰 생성
-        String access = jwtUtil.createJwt("access", email, role, 600000L);
+        String access = jwtUtil.createJwt("access", email, role, 1800000L);
         String refresh = jwtUtil.createJwt("refresh", email, role, 86400000L);
         log.info("Generated access and refresh tokens for user: {}", email);
